splitting not needed since not using interface
